LP_9_08_js_i18n  Добавлена интернационализация в Javascript  на стороне сервера формируется i18n JavaScript массив с значениями, который затем используется для интернационализации в браузере     в модальном окне заголовок подменяется через $('#modalTitle').html(..title)

+++ src/main/webapp/resources/js/topjava.common.js
@@ -16,21 +16,25 @@ function makeEditable(ctx) {
 }
 
 function add() {
+    $("#modalTitle").html(i18n["addTitle"]);
     form.find(":input").val("");
     $("#editRow").modal();
 }
 
 function deleteRow(id) {
-    $.ajax({
-        url: context.ajaxUrl + id,
-        type: "DELETE"
-    }).done(function () {
-        context.updateTable();
-        successNoty("пользователь удален!")
-    });
+    if(confirm(i18n["common.confirm"])) {
+        $.ajax({
+            url: context.ajaxUrl + id,
+            type: "DELETE"
+        }).done(function () {
+            context.updateTable();
+            successNoty("common.deleted");
+        });
+    }
 }
 
 function updateRow(id){
+    $("#modalTitle").html(i18n["editTitle"]);
     $.get(
         context.ajaxUrl+id, function(data){
             $.each(data,function(key,value){
@@ -59,7 +63,7 @@ function save() {
     }).done(function () {
         $("#editRow").modal("hide");
         context.updateTable();
-        successNoty("Пользователь сохранен!");
+        successNoty("common.saved");
     });
 }
 
@@ -72,10 +76,10 @@ function closeNoty(){
     }
 }
 
-function successNoty(text){
+function successNoty(key){
     closeNoty();
     new Noty({
-        text: "<span class='fa fa-lg fa-check'></span> &nbsp;" + text,
+        text: "<span class='fa fa-lg fa-check'></span> &nbsp;" + i18n[key],
         type: 'success',
         layout: "bottomRight",
         timeout: 1000
@@ -85,7 +89,7 @@ function successNoty(text){
 function failNoty(jqXHR){
     closeNoty();
     failedNote = new Noty({
-        text: "<span class='fa fa-lg fa-exclamation-circle'></span> &nbsp;Error status: " + jqXHR.status + (jqXHR.responseJSON ? "<br>" + jqXHR.responseJSON : ""),
+        text: "<span class='fa fa-lg fa-exclamation-circle'></span> &nbsp; " +i18n["common.errorStatus"]+ jqXHR.status + (jqXHR.responseJSON ? "<br>" + jqXHR.responseJSON : ""),
         type: "error",
         layout: "bottomRight"
     }).show();
